Hashtable instead of ifs for converters.
