+++ nanoFramework.Json/Converters/BoolConverter.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace nanoFramework.Json.Converters
+{
+    internal sealed class BoolConverter : IConverter
+    {
+        public string ToJson(object value)
+        {
+            return (bool)value ? "true" : "false";
+        }
+
+        public object ToType(object value)
+        {
+            return Convert.ToBoolean(Convert.ToByte(value.ToString()));
+        }
+    }
+}
