+++ nanoFramework.Json/Converters/DoubleConverter.cs
@@ -1,0 +1,22 @@
+﻿using System;
+
+namespace nanoFramework.Json.Converters
+{
+    internal sealed class DoubleConverter : IConverter
+    {
+        public string ToJson(object value)
+        {
+            if (double.IsNaN((double)value))
+            {
+                return "null";
+            }
+
+            return value.ToString();
+        }
+
+        public object ToType(object value)
+        {
+            return Convert.ToDouble(value.ToString());
+        }
+    }
+}
