+++ nanoFramework.Json/Converters/IConverter.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Text;
+
+namespace nanoFramework.Json.Converters
+{
+    internal interface IConverter
+    {
+        object ToType(object value);
+        string ToJson(object value);
+    }
+}
