+++ nanoFramework.Json/Converters/IntConverter.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace nanoFramework.Json.Converters
+{
+    internal sealed class IntConverter : IConverter
+    {
+        public string ToJson(object value)
+        {
+            return value.ToString();
+        }
+
+        public object ToType(object value)
+        {
+            return Convert.ToInt32(value.ToString());
+        }
+    }
+}
