+++ nanoFramework.Json/Converters/SByteConverter.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace nanoFramework.Json.Converters
+{
+    internal sealed class SByteConverter : IConverter
+    {
+        public string ToJson(object value)
+        {
+            return value.ToString();
+        }
+
+        public object ToType(object value)
+        {
+            return Convert.ToSByte(value.ToString());
+        }
+    }
+}
