+++ nanoFramework.Json/Converters/ShortConverter.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace nanoFramework.Json.Converters
+{
+    internal sealed class ShortConverter : IConverter
+    {
+        public string ToJson(object value)
+        {
+            return value.ToString();
+        }
+
+        public object ToType(object value)
+        {
+            return Convert.ToInt16(value.ToString());
+        }
+    }
+}
