+++ nanoFramework.Json/Converters/TimeSpanConverter.cs
@@ -1,0 +1,15 @@
+﻿namespace nanoFramework.Json.Converters
+{
+    internal sealed class TimeSpanConverter : IConverter
+    {
+        public string ToJson(object value)
+        {
+            return "\"" + value.ToString() + "\"";
+        }
+
+        public object ToType(object value)
+        {
+            return TimeSpanExtensions.ConvertFromString(value.ToString());
+        }
+    }
+}
